Add more shared code
