+++ tools/c7n_mailer/c7n_mailer/azure/azure_queue_processor.py
@@ -22,7 +22,7 @@ import traceback
 import zlib
 
 import six
-from c7n_mailer.azure.azure_smtp_delivery import AzureSmtpDelivery
+from c7n_mailer.smtp_delivery import SmtpDelivery
 from c7n_mailer.azure.sendgrid_delivery import SendGridDelivery
 
 try:
@@ -114,9 +114,9 @@ class MailerAzureQueueProcessor(object):
             sendgrid_messages = sendgrid_delivery.get_to_addrs_sendgrid_messages_map(queue_message)
 
             if 'smtp_server' in self.config:
-                smtp_delivery = AzureSmtpDelivery(config=self.config,
-                                                  session=self.session,
-                                                  logger=self.logger)
+                smtp_delivery = SmtpDelivery(config=self.config,
+                                             session=self.session,
+                                             logger=self.logger)
                 for to_addrs, message in six.iteritems(sendgrid_messages):
                     smtp_delivery.send_message(message=message, to_addrs=list(to_addrs))
             else:
