+++ tools/c7n_mailer/tests/test_smtp_delivery.py
@@ -1,7 +1,5 @@
 import unittest
 
-from c7n_mailer.aws.aws_smtp_delivery import AwsSmtpDelivery
-from c7n_mailer.azure.azure_smtp_delivery import AzureSmtpDelivery
 from c7n_mailer.smtp_delivery import SmtpDelivery
 from mock import patch, call, MagicMock
 
@@ -10,23 +8,30 @@ class SmtpDeliveryTest(unittest.TestCase):
 
     @patch('smtplib.SMTP')
     def test_no_ssl(self, mock_smtp):
-        d = SmtpDelivery(smtp_server='server',
-                         smtp_port=25,
-                         smtp_ssl=False,
-                         smtp_username=None,
-                         smtp_password=None)
+        config = {
+            'smtp_server': 'server',
+            'smtp_port': 25,
+            'smtp_ssl': False,
+            'smtp_username': None,
+            'smtp_password': None
+        }
+        d = SmtpDelivery(config, MagicMock(), MagicMock())
         del d
 
         mock_smtp.assert_has_calls([call('server', 25),
                                     call().quit()])
 
+    @patch('c7n_mailer.utils.kms_decrypt', return_value='password')
     @patch('smtplib.SMTP')
-    def test_no_ssl_with_credentials(self, mock_smtp):
-        d = SmtpDelivery(smtp_server='server',
-                         smtp_port=25,
-                         smtp_ssl=False,
-                         smtp_username='username',
-                         smtp_password='password')
+    def test_no_ssl_with_credentials(self, mock_smtp, kms_mock):
+        config = {
+            'smtp_server': 'server',
+            'smtp_port': 25,
+            'smtp_ssl': False,
+            'smtp_username': 'username',
+            'smtp_password': 'test'
+        }
+        d = SmtpDelivery(config, MagicMock(), MagicMock())
         del d
 
         mock_smtp.assert_has_calls([call('server', 25),
@@ -35,11 +40,14 @@ class SmtpDeliveryTest(unittest.TestCase):
 
     @patch('smtplib.SMTP')
     def test_with_ssl(self, mock_smtp):
-        d = SmtpDelivery(smtp_server='server',
-                         smtp_port=25,
-                         smtp_ssl=True,
-                         smtp_username=None,
-                         smtp_password=None)
+        config = {
+            'smtp_server': 'server',
+            'smtp_port': 25,
+            'smtp_ssl': True,
+            'smtp_username': None,
+            'smtp_password': None
+        }
+        d = SmtpDelivery(config, MagicMock(), MagicMock())
         del d
 
         mock_smtp.assert_has_calls([call('server', 25),
@@ -47,13 +55,17 @@ class SmtpDeliveryTest(unittest.TestCase):
                                     call().ehlo(),
                                     call().quit()])
 
+    @patch('c7n_mailer.utils.kms_decrypt', return_value='password')
     @patch('smtplib.SMTP')
-    def test_with_ssl_and_credentials(self, mock_smtp):
-        d = SmtpDelivery(smtp_server='server',
-                         smtp_port=25,
-                         smtp_ssl=True,
-                         smtp_username='username',
-                         smtp_password='password')
+    def test_with_ssl_and_credentials(self, mock_smtp, kms_mock):
+        config = {
+            'smtp_server': 'server',
+            'smtp_port': 25,
+            'smtp_ssl': True,
+            'smtp_username': 'username',
+            'smtp_password': 'test'
+        }
+        d = SmtpDelivery(config, MagicMock(), MagicMock())
         del d
 
         mock_smtp.assert_has_calls([call('server', 25),
@@ -64,11 +76,14 @@ class SmtpDeliveryTest(unittest.TestCase):
 
     @patch('smtplib.SMTP')
     def test_send_message(self, mock_smtp):
-        d = SmtpDelivery(smtp_server='server',
-                         smtp_port=25,
-                         smtp_ssl=False,
-                         smtp_username=None,
-                         smtp_password=None)
+        config = {
+            'smtp_server': 'server',
+            'smtp_port': 25,
+            'smtp_ssl': False,
+            'smtp_username': None,
+            'smtp_password': None
+        }
+        d = SmtpDelivery(config, MagicMock(), MagicMock())
         message_mock = MagicMock()
         message_mock.__getitem__.side_effect = lambda x: 't@test.com' if x == 'From' else None
         message_mock.as_string.return_value = 'mock_text'
@@ -78,41 +93,4 @@ class SmtpDeliveryTest(unittest.TestCase):
 
         mock_smtp.assert_has_calls([call('server', 25),
                                     call().sendmail('t@test.com', ['test1@test.com'], 'mock_text'),
-                                    call().quit()])
-
-    @patch('smtplib.SMTP')
-    def test_azure_smtp_delivery(self, mock_smtp):
-        config = {
-            'smtp_server': 'server',
-            'smtp_port': 25,
-            'smtp_ssl': True,
-            'smtp_username': 'username',
-            'smtp_password': 'password'
-        }
-        a = AzureSmtpDelivery(config, None, None)
-        del a
-
-        mock_smtp.assert_has_calls([call('server', 25),
-                                    call().starttls(),
-                                    call().ehlo(),
-                                    call().login('username', 'password'),
-                                    call().quit()])
-
-    @patch('c7n_mailer.utils.kms_decrypt', return_value='password')
-    @patch('smtplib.SMTP')
-    def test_aws_smtp_delivery(self, mock_smtp, kms_mock):
-        config = {
-            'smtp_server': 'server',
-            'smtp_port': 25,
-            'smtp_ssl': True,
-            'smtp_username': 'username',
-            'smtp_password': 'pass'
-        }
-        a = AwsSmtpDelivery(config, MagicMock(), MagicMock())
-        del a
-
-        mock_smtp.assert_has_calls([call('server', 25),
-                                    call().starttls(),
-                                    call().ehlo(),
-                                    call().login('username', 'password'),
                                     call().quit()])
